Bboxes work for GridElasticTransform
